Support reading bytes directly from duck-typed readables
There are many use cases when the source file does not reside on a
traditional filesystem.  These are already supported because get_bytes
accepts a byte array.

However, in cases where it is advantageous to only read the necessary
header bytes (e.g. across a network), currently filetype expects the
user to know that only 262 bytes are necessary, generally requiring
this constant to be defined in userland code.

Python transparently supports the ability to *attempt* to read the
first 262 bytes from whatever object is passed to it via duck typing.
This means that an S3 StreamingBody, for instance, which reads out
bytes, could be passed directly to filetype.guess/filetype.match,
without needing to read out the first 262 bytes manually in userland
code.

A minor convenience, but a win for usability and code cleanliness!